+++ Customer/src/main/java/com/example/customer/entity/ProductEntity.java
@@ -19,9 +19,9 @@ public class ProductEntity {
 
     private String name;
 
-    private double original_price;
+    private Long original_price;
 
-    private double price;
+    private Long price;
 
     private String description;
 
@@ -33,7 +33,7 @@ public class ProductEntity {
 
     private double overall_rating;
 
-    private double discount;
+    private Long discount;
 
     private String image1;
